--- js/functions.js
@@ -1,38 +1,0 @@
-const checkString = (stringChars, lengthMax) =>
-  stringChars.length <= lengthMax;
-
-checkString('проверяемая строка', 20);
-checkString('проверяемая строка', 18);
-checkString('проверяемая строка', 10);
-
-function checkPalindrom(rawString) {
-  const string = rawString.replaceAll(' ', '').toLowerCase();
-  for (let i = 0; i < string.length / 2; i++) {
-    if (string.at(i) !== string.at(-i - 1)) {
-      return false;
-    }
-  }
-  return true;
-}
-
-checkPalindrom('топот');
-checkPalindrom('ДовОд');
-checkPalindrom('Кекс');
-checkPalindrom('Лёша на полке клопа нашёл ');
-
-function extractNumber(arg) {
-  const string = arg.toString();
-  let result = '';
-  for (let i = 0; i < string.length; i++) {
-    if (!Number.isNaN(parseInt(string[i], 10))) {
-      result += string[i];
-    }
-  }
-  return parseInt(result,10);
-}
-
-extractNumber('2023 год');
-extractNumber('ECMAScript 2022');
-extractNumber('1 кефир, 0.5 батона');
-extractNumber('агент 007');
-extractNumber('а я томат');
